check if input configuration exists, otherwise keep gauge_start set to random; handle cases where no configurations an/or out_0 does not exist

+++ input/templates/update_replicas.sh
@@ -22,19 +22,28 @@ fi
 
 for (( i=0; i<$r; i+=1 )); do
 
-    gsfile=$(ls Rep_${i}/run1* -t | head -1)
-    gsfile=${gsfile#"Rep_${i}/"}
-    # get the name of the last Robbins-Monro trajectory
-    # then increment by one to get the next Robbins-Monro weight.
-    RM_NUM=$(grep 'Robbins Monro sequence #' Rep_0/out_0 | tail -n 1 | grep -oP '(?<=#).*?(?=:)')
-    de=$(grep "LLR Delta S" Rep_$i/out_0 | grep -o -E '[0-9]+(\.[0-9]+)'| tail -n 1)
-    E=$(grep "a_rho(0," Rep_$i/out_0 | tail -1 | grep -o -E '[0-9]+(\.[0-9]+)'| head -n 1)
-    A=$(grep "a_rho(0," Rep_$i/out_0 | tail -1 | grep -o -E '[0-9]+(\.[0-9]+)'| tail -n 1)
+    # find all configurations in relevant directory
+    confs=$(find ./Rep_${i} -name run1*)
+    if [[ -n $confs ]]; then
+        gsfile=$(ls Rep_${i}/run1* -t | head -1)
+    fi
+    
+    if [[ -f Rep_$i/out_0 ]]; then 
+        RM_NUM=$(grep 'Robbins Monro sequence #' Rep_0/out_0 | tail -n 1 | grep -oP '(?<=#).*?(?=:)')
+        de=$(grep "LLR Delta S" Rep_$i/out_0 | grep -o -E '[0-9]+(\.[0-9]+)'| tail -n 1)
+        E=$(grep "a_rho(0," Rep_$i/out_0 | tail -1 | grep -o -E '[0-9]+(\.[0-9]+)'| head -n 1)
+        A=$(grep "a_rho(0," Rep_$i/out_0 | tail -1 | grep -o -E '[0-9]+(\.[0-9]+)'| tail -n 1)
+    fi
 
     sed -i "/rlx_seed/c\rlx_seed = ${RANDOM}"         Rep_${i}/$FILEA
-    sed -i "/gauge start/c\gauge start = ${gsfile}"   Rep_${i}/$FILEA
-    sed -i "/llr:dS/c\llr:dS = ${de}"                 Rep_${i}/$FILEA
     # For the following quantities first test if we were able to read the required information from the logs
+    if [[ -n ${gsfile} ]]; then
+        gsfile=${gsfile#"Rep_${i}/"}
+        sed -i "/gauge start/c\gauge start = ${gsfile}" Rep_${i}/$FILEA
+    fi
+    if [[ -n ${de} ]]; then
+        sed -i "/llr:dS/c\llr:dS = ${de}"             Rep_${i}/$FILEA
+    fi
     if [[ -n $E ]]; then
         sed -i "/llr:S0/c\llr:S0 = $E"                Rep_${i}/$FILEA
     fi
@@ -42,20 +51,23 @@ for (( i=0; i<$r; i+=1 )); do
         sed -i "/llr:starta/c\llr:starta = ${A}"      Rep_${i}/$FILEA
     fi
     if [[ -n $RM_NUM ]]; then
+        # increment by one to get the next Robbins-Monro weight.
         sed -i "/llr:it =/c\llr:it = $(($RM_NUM+1))"  Rep_${i}/$FILEA
     fi
 
     # remove old configuration files
-    # check if all files are of the same size
-    same_size=$(readlink -f Rep_${i}/run1* | xargs du | awk '{print $1}' | uniq -u | wc -l)
-    # same_size is equal to zero if all files are of the same size
-    # only if this is the case this script will remove old configurations
-    if [ $same_size -eq 0 ]; then
-        readlink -f Rep_${i}/run1* >> tmp_list
-        if [ $(wc -l tmp_list | awk '{print $1}') -gt 1 ]; then
-            cat tmp_list | sort -V | head -n -1 | xargs rm
+    if [[ -n $confs ]]; then
+        # check if all files are of the same size
+        same_size=$(readlink -f Rep_${i}/run1* | xargs du | awk '{print $1}' | uniq -u | wc -l)
+        # same_size is equal to zero if all files are of the same size
+        # only if this is the case this script will remove old configurations
+        if [ $same_size -eq 0 ]; then
+            readlink -f Rep_${i}/run1* >> tmp_list
+            if [ $(wc -l tmp_list | awk '{print $1}') -gt 1 ]; then
+                cat tmp_list | sort -V | head -n -1 | xargs rm
+            fi
+        rm tmp_list
         fi
-	rm tmp_list
     fi
 
 done
